requirements.txt
Flask>=2.3.0,<3.0.0
pymysql>=1.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
Flask-Mail>=0.9.0
werkzeug>=2.3.0,<3.0.0
xlsxwriter>=3.0.0
openpyxl>=3.0.0
apscheduler>=3.10.0
cryptography>=41.0.0